Move mystrerror to a C file
This is to resolve errors on CentOS 7.
While there:
 * moved the other C file, my_aligned_alloc.c, to the util directory
 * removed bswap.h and sys_basic.h since these are no longer needed
   (they were used for an endianness check but PR #17648 addressed it
    in a different way).
 * renamed files.cpp/files.h to filesystem.cpp/filesystem.h to
   avoid having 2 files with the same name.

---
