--- compiler/next/lib/util/bswap.h
@@ -1,1 +1,0 @@
-../../../../runtime/include/qio/bswap.h
