+++ compiler/next/lib/util/filesystem.h
@@ -17,14 +17,31 @@
  * limitations under the License.
  */
 
-#ifndef MYSTRERROR_H
-#define MYSTRERROR_H
+#ifndef CHPL_UTIL_FILESYSTEM_H
+#define CHPL_UTIL_FILESYSTEM_H
 
+#include "chpl/queries/ErrorMessage.h"
+
+#include <cstdio>
 #include <string>
 
 namespace chpl {
 
-std::string my_strerror(int errno_);
+/**
+  Open a file. If the open failed, return nullptr and set errorOut.
+ */
+FILE* openfile(const char* path, const char* mode, ErrorMessage& errorOut);
+
+/**
+  Close a file. If the close failed, return false and set errorOut.
+ */
+bool closefile(FILE* fp, const char* path, ErrorMessage& errorOut);
+
+/**
+  Reads the contents of a file into a string.
+  If something failed, returns false and sets errorOut.
+ */
+bool readfile(const char* path, std::string& strOut, ErrorMessage& errorOut);
 
 } // end namespace chpl
