+++ compiler/next/lib/util/my_strerror_r.c
@@ -17,32 +17,14 @@
  * limitations under the License.
  */
 
-#ifndef CHPL_UTIL_FILES_H
-#define CHPL_UTIL_FILES_H
+// make sure to get the XSI/POSIX.1-2001 strerror_r
+#define _POSIX_C_SOURCE 200112L
+#undef _GNU_SOURCE
 
-#include "chpl/queries/ErrorMessage.h"
+#include <string.h>
 
-#include <cstdio>
-#include <string>
+#include "my_strerror_r.h"
 
-namespace chpl {
-
-/**
-  Open a file. If the open failed, return nullptr and set errorOut.
- */
-FILE* openfile(const char* path, const char* mode, ErrorMessage& errorOut);
-
-/**
-  Close a file. If the close failed, return false and set errorOut.
- */
-bool closefile(FILE* fp, const char* path, ErrorMessage& errorOut);
-
-/**
-  Reads the contents of a file into a string.
-  If something failed, returns false and sets errorOut.
- */
-bool readfile(const char* path, std::string& strOut, ErrorMessage& errorOut);
-
-} // end namespace chpl
-
-#endif
+int my_strerror_r(int errnum, char* strerrbuf, size_t buflen) {
+  return strerror_r(errnum, strerrbuf, buflen);
+}
