--- compiler/next/lib/util/sys_basic.h
@@ -1,1 +1,0 @@
-../../../../runtime/include/sys_basic.h
